requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-chroma
chromadb